IoT/kasa_devices/plugs/container_influxdb/requirements.txt
python-kasa
asyncio
paho-mqtt==1.6.1